Add files via upload
food intake-analysis tool (11/07/2024):
This ImageJ-script calculates the percentage of ingested colored yeast relative to the body surface in Drosophila larva. Images of the larva should be photographed by dark-field illumination. To analyze images should be organized in folders. Open one image. After starting the script, you need to enter the number images in the respective folder and adjust threshold (“Threshold” dialog window set to “Default” and “Red”; activate “Dark background” and “Don’t reset range”). Only press the “Adjust” bottom in the “Action Required” window. After this step, the script analyzes all images in the folder and creates a “Processed” folder which include the images for the detected body surface and ingested yeast.